NIFI-9502 Move ExtensionManifest parser and impl from nifi registry into common module
