+++ nifi-manifest/nifi-extension-manifest-parser/pom.xml
@@ -17,18 +17,23 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.apache.nifi</groupId>
-        <artifactId>nifi</artifactId>
+        <artifactId>nifi-manifest</artifactId>
         <version>1.16.0-SNAPSHOT</version>
     </parent>
-    <artifactId>nifi-manifest</artifactId>
-    <packaging>pom</packaging>
+    <artifactId>nifi-extension-manifest-parser</artifactId>
+    <packaging>jar</packaging>
 
-    <modules>
-	<module>nifi-extension-manifest-model</module>
-        <module>nifi-extension-manifest-parser</module>
-        <module>nifi-runtime-manifest-core</module>
-        <module>nifi-runtime-manifest</module>
-        <module>nifi-runtime-manifest-test</module>
-    </modules>
+    <dependencies>
+        <dependency>
+            <groupId>org.apache.nifi</groupId>
+            <artifactId>nifi-extension-manifest-model</artifactId>
+            <version>1.16.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>com.fasterxml.jackson.dataformat</groupId>
+            <artifactId>jackson-dataformat-xml</artifactId>
+            <version>${jackson.version}</version>
+        </dependency>
+    </dependencies>
 
 </project>
