+++ nifi-manifest/nifi-extension-manifest-parser/src/test/resources/manifests/extension-manifest-dbcp-service-nar.xml
