+++ nifi-manifest/nifi-extension-manifest-parser/src/test/resources/manifests/extension-manifest-kafka-2-6-nar.xml
