+++ nifi-manifest/nifi-extension-manifest-parser/src/test/resources/manifests/extension-manifest-missing-sys-api.xml
