+++ nifi-manifest/nifi-runtime-manifest-core/pom.xml
@@ -30,13 +30,18 @@
             <version>1.16.0-SNAPSHOT</version>
         </dependency>
         <dependency>
-            <groupId>org.apache.nifi.registry</groupId>
-            <artifactId>nifi-registry-bundle-utils</artifactId>
+            <groupId>org.apache.nifi</groupId>
+            <artifactId>nifi-api</artifactId>
             <version>1.16.0-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
             <artifactId>nifi-extension-manifest-model</artifactId>
+            <version>1.16.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.nifi</groupId>
+            <artifactId>nifi-extension-manifest-parser</artifactId>
             <version>1.16.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
