+++ nifi-registry/nifi-registry-core/nifi-registry-bundle-utils/pom.xml
@@ -27,9 +27,9 @@
             <version>1.16.0-SNAPSHOT</version>
         </dependency>
         <dependency>
-            <groupId>com.fasterxml.jackson.dataformat</groupId>
-            <artifactId>jackson-dataformat-xml</artifactId>
-            <version>${jackson.version}</version>
+            <groupId>org.apache.nifi</groupId>
+            <artifactId>nifi-extension-manifest-parser</artifactId>
+            <version>1.16.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
